requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0
django-allauth==0.40.0
django-debug-toolbar==2.1
django-decouple==2.1
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.11.0
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2
